Move EKS clusterbuilder under capa
