+++ pkg/clusterbuilder/providers/capa/values/managed-cluster_values.yaml
