+++ pkg/clusterbuilder/providers/capa/values/managed-default-apps_values.yaml
